Django-Telex-Integtation/requirements.txt
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.19
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
